requirements.txt
opencv-python>=4.5.3.56
Flask>=2.0.1
Flask-MySQLdb>=0.2.0
numpy>=1.21.2
pandas>=1.3.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2>=3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
PyYAML==5.4.1